CChildView: Change locking.
The lock now uses an int. Setting the locked flag has been replaced by
incrementing the lock variable, and clearing it by decrementing it. The
lock is now held for shorter time periods, and locking was added to
OnValueEditUpdate. In OnHexColorUpdate, also move the UpdateBitmaps call
inside the fChanged block.